Refactor ActiveFilters to use Tag component
- Move getActiveFilters query to data.remote.ts for consistency
- Use Tag component with removable prop for tag-type filters
- Add bold type labels and svelte-orange hover color on close buttons
